+++ client-v2/src/test/java/com/clickhouse/client/query/QueryTests.java
@@ -18,6 +18,7 @@ import com.clickhouse.client.api.data_formats.NativeFormatReader;
 import com.clickhouse.client.api.data_formats.RowBinaryFormatReader;
 import com.clickhouse.client.api.data_formats.RowBinaryWithNamesAndTypesFormatReader;
 import com.clickhouse.client.api.data_formats.RowBinaryWithNamesFormatReader;
+import com.clickhouse.client.api.insert.InsertSettings;
 import com.clickhouse.client.api.metadata.TableSchema;
 import com.clickhouse.client.api.query.NullValueException;
 import com.clickhouse.client.api.query.QueryResponse;
@@ -31,15 +32,18 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -87,8 +91,7 @@ public class QueryTests extends BaseIntegrationTest {
         QuerySettings settings = new QuerySettings()
                 .setFormat(ClickHouseFormat.TabSeparated);
 
-        Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE,
-                Collections.emptyMap(), settings);
+        Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE, settings);
 
         try {
             BufferedReader br = new BufferedReader(new InputStreamReader(response.get().getInputStream()));
@@ -118,7 +121,7 @@ public class QueryTests extends BaseIntegrationTest {
 
         Iterator<Boolean> expectedIterator = expected.iterator();
         for (QuerySettings settings : settingsList) {
-            Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE, null, settings);
+            Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE, settings);
             QueryResponse queryResponse = response.get();
             ArrayList<JsonNode> records = new ArrayList<>();
             final ObjectMapper objectMapper = new ObjectMapper();
@@ -155,7 +158,7 @@ public class QueryTests extends BaseIntegrationTest {
         List<Map<String, Object>> data = prepareDataSet(DATASET_TABLE + "_" + format.name(), DATASET_COLUMNS,
                 DATASET_VALUE_GENERATORS, rows);
         QuerySettings settings = new QuerySettings().setFormat(format);
-        Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE + "_" + format.name(), null, settings);
+        Future<QueryResponse> response = client.query("SELECT * FROM " + DATASET_TABLE + "_" + format.name(), settings);
         QueryResponse queryResponse = response.get();
 
         Map<String, Object> record = new HashMap<>();
@@ -203,7 +206,7 @@ public class QueryTests extends BaseIntegrationTest {
         List<Map<String, Object>> data = prepareDataSet(table, DATASET_COLUMNS,
                 DATASET_VALUE_GENERATORS, 10);
         QuerySettings settings = new QuerySettings().setFormat(ClickHouseFormat.RowBinaryWithNamesAndTypes);
-        Future<QueryResponse> response = client.query("SELECT col1, col3, hostname() as host FROM " + table, null, settings);
+        Future<QueryResponse> response = client.query("SELECT col1, col3, hostname() as host FROM " + table, settings);
         QueryResponse queryResponse = response.get();
 
         TableSchema schema = new TableSchema();
@@ -233,7 +236,7 @@ public class QueryTests extends BaseIntegrationTest {
         final int rows = 10;
         List<Map<String, Object>> data = prepareDataSet(table, DATASET_COLUMNS, DATASET_VALUE_GENERATORS, rows);
         QuerySettings settings = new QuerySettings().setFormat(ClickHouseFormat.RowBinaryWithNamesAndTypes);
-        Future<QueryResponse> response = client.query("SELECT col1, col3, hostname() as host FROM " + table, null, settings);
+        Future<QueryResponse> response = client.query("SELECT col1, col3, hostname() as host FROM " + table, settings);
 
         QueryResponse queryResponse = response.get();
         TableSchema schema = new TableSchema();
@@ -275,7 +278,7 @@ public class QueryTests extends BaseIntegrationTest {
         List<Map<String, Object>> data = prepareDataSet(table, ARRAY_COLUMNS, ARRAY_VALUE_GENERATORS, rows);
 
         QuerySettings settings = new QuerySettings().setFormat(ClickHouseFormat.RowBinaryWithNamesAndTypes);
-        Future<QueryResponse> response = client.query("SELECT * FROM " + table, null, settings);
+        Future<QueryResponse> response = client.query("SELECT * FROM " + table, settings);
         TableSchema schema = client.getTableSchema(table);
 
         QueryResponse queryResponse = response.get();
@@ -320,7 +323,7 @@ public class QueryTests extends BaseIntegrationTest {
         List<Map<String, Object>> data = prepareDataSet(table, MAP_COLUMNS, MAP_VALUE_GENERATORS, rows);
 
         QuerySettings settings = new QuerySettings().setFormat(ClickHouseFormat.RowBinaryWithNamesAndTypes);
-        Future<QueryResponse> response = client.query("SELECT * FROM " + table, null, settings);
+        Future<QueryResponse> response = client.query("SELECT * FROM " + table, settings);
         TableSchema schema = client.getTableSchema(table);
 
         QueryResponse queryResponse = response.get();
@@ -364,7 +367,7 @@ public class QueryTests extends BaseIntegrationTest {
         List<Map<String, Object>> data = prepareDataSet(table, NULL_DATASET_COLUMNS, NULL_DATASET_VALUE_GENERATORS, 1);
 
         QuerySettings settings = new QuerySettings().setFormat(ClickHouseFormat.RowBinaryWithNamesAndTypes);
-        Future<QueryResponse> response = client.query("SELECT * FROM " + table, null, settings);
+        Future<QueryResponse> response = client.query("SELECT * FROM " + table, settings);
         TableSchema schema = client.getTableSchema(table);
 
         QueryResponse queryResponse = response.get();
@@ -839,7 +842,7 @@ public class QueryTests extends BaseIntegrationTest {
         }
         selectStmtBuilder.setLength(selectStmtBuilder.length() - 2);
         selectStmtBuilder.append(" FROM ").append(table);
-        Future<QueryResponse> response = client.query(selectStmtBuilder.toString(), null, settings);
+        Future<QueryResponse> response = client.query(selectStmtBuilder.toString(), settings);
         TableSchema schema = client.getTableSchema(table);
 
         try {
@@ -869,8 +872,7 @@ public class QueryTests extends BaseIntegrationTest {
         QuerySettings settings = new QuerySettings()
                 .setFormat(ClickHouseFormat.TabSeparated);
 
-        QueryResponse response = client.query("SELECT * FROM " + DATASET_TABLE + " LIMIT 3",
-                Collections.emptyMap(), settings).get();
+        QueryResponse response = client.query("SELECT * FROM " + DATASET_TABLE + " LIMIT 3", settings).get();
 
         // Stats should be available after the query is done
         OperationStatistics stats = response.getOperationStatistics();
@@ -893,8 +895,7 @@ public class QueryTests extends BaseIntegrationTest {
             insertStmtBuilder.setLength(insertStmtBuilder.length() - 2);
             insertStmtBuilder.append("), ");
         }
-        response = client.query(insertStmtBuilder.toString(),
-                Collections.emptyMap(), settings).get();
+        response = client.query(insertStmtBuilder.toString(), settings).get();
 
         serverStats = response.getOperationStatistics().serverStatistics;
         System.out.println("Server stats: " + serverStats);
@@ -1033,5 +1034,32 @@ public class QueryTests extends BaseIntegrationTest {
         }
         sb.setLength(sb.length() - 2);
         sb.append(']');
+    }
+
+    @Test(groups = {"integration"})
+    public void testQueryParams() throws Exception {
+        final String table = "query_params_test_table";
+
+        client.query("DROP TABLE IF EXISTS default." + table, null).get();
+        client.query("CREATE TABLE default." + table + " (col1 UInt32, col2 String) ENGINE = MergeTree ORDER BY tuple()").get();
+
+        ByteArrayOutputStream insertData = new ByteArrayOutputStream();
+        try (BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(insertData))) {
+            writer.write("1\t'one'\n");
+            writer.write("2\t'two'\n");
+            writer.write("3\t'three'\n");
+        }
+        InsertSettings insertSettings = new InsertSettings().setFormat(ClickHouseFormat.TabSeparated);
+        client.insert(table, new ByteArrayInputStream(insertData.toByteArray()), insertSettings).get();
+
+        QuerySettings querySettings = new QuerySettings().setFormat(ClickHouseFormat.TabSeparatedWithNamesAndTypes.name());
+        Map<String, Object> queryParams = new HashMap<>();
+        queryParams.put("param1", 2);
+        QueryResponse queryResponse =
+                client.query("SELECT * FROM " + table + " WHERE col1 >= {param1:UInt32}", queryParams, querySettings).get();
+
+        try (BufferedReader responseBody = new BufferedReader(new InputStreamReader(queryResponse.getInputStream()))) {
+            responseBody.lines().forEach(System.out::println);
+        }
     }
 }