+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Latest 
 
+## 0.6.0-patch5
+
+### Bug Fixes
+- Roles (memorized by client) can be reset with 'SET ROLE NONE' query (https://github.com/ClickHouse/clickhouse-java/issues/1647)
+- MaxBufferSize can be greater than internal MAX value now
+- Updated example project to use the latest version of the client
+
 ## 0.6.0-patch4
 
 ### New Features